#345 Include Unit Tests in project Mocha + Chai + unit test utilsService.js, SubmissionInfo.vue and store + refactoring store: split files getters, mutations and actions to allow unit testing
