+++ frontend/package.json
@@ -4,7 +4,9 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "build": "vue-cli-service build",
+    "test": "vue-cli-service test:unit --watch",
+    "testNoWatch": "vue-cli-service test:unit"
   },
   "dependencies": {
     "@coreui/coreui": "^2.0.15",
@@ -27,15 +29,16 @@
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.0.5",
     "@vue/cli-plugin-eslint": "^3.1.5",
+    "@vue/cli-plugin-unit-mocha": "^3.2.0",
     "@vue/cli-service": "^3.0.5",
     "@vue/eslint-config-airbnb": "^4.0.0",
-    "@vue/test-utils": "^1.0.0-beta.20",
     "babel-eslint": "^10.0.1",
-    "chai": "^4.1.2",
+    "chai": "^4.2.0",
     "eslint": "^5.8.0",
     "eslint-plugin-vue": "^5.0.0-0",
     "node-sass": "^4.9.0",
     "sass-loader": "^7.0.1",
+    "sinon": "^7.2.2",
     "vue-template-compiler": "^2.5.17"
   }
 }
