+++ frontend/src/components/common/SubmissionInfo.vue
@@ -28,7 +28,6 @@
         </div>
       </b-card>
     </form>
-
     </div>
 </template>
 
@@ -38,9 +37,6 @@ import fieldGenerator from '@/components/common/form-components/fieldGenerator'
 import labels from '@/components/art7/dataDefinitions/labels'
 
 export default {
-
-	name: 'Tab1',
-
 	props: {
 		info: Object,
 		tabs: Object
@@ -63,7 +59,3 @@ export default {
 
 }
 </script>
-
-<style lang="css" scoped>
-
-</style>
