+++ frontend/tests/unit/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+	env: {
+	  mocha: true,
+	},
+  };
+
